finshed some product actions

+++ Löwen.Application/Features/DiscountFeature/Commands/UpdateDiscount/UpdateDiscountCommand.cs
@@ -1,4 +1,4 @@
 ﻿namespace Löwen.Application.Features.DiscountFeature.Commands.UpdateDiscount;
 
 public record UpdateDiscountCommand(string Id,string? Name, DiscountType? DiscountType,
-    decimal? DiscountValue, DateTime? StartDate, DateTime? EndDate, bool? IsActive) : ICommand;
+    double? DiscountValue, DateTime? StartDate, DateTime? EndDate, bool? IsActive) : ICommand;
