+++ Löwen.Application/Features/DiscountFeature/Queries/Response/DiscountResponse.cs
@@ -12,7 +12,7 @@ namespace Löwen.Application.Features.DiscountFeature.Queries.Response
         public Guid Id { get; set; }
         public required string Name { get; set; } 
         public DiscountType DiscountType { get; set; }
-        public decimal? DiscountValue { get; set; }
+        public double? DiscountValue { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
         public bool IsActive { get; set; }
